# api/openapi.json
{
  "components": {
    "schemas": {
      "Body_batch_channel_api_batch_channel_post": {
        "properties": {
          "audio": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio"
          },
          "brand_name": {
            "title": "Brand Name",
            "type": "string"
          },
          "channel_archive": {
            "description": "Zip of subfolders containing images",
            "format": "binary",
            "title": "Channel Archive",
            "type": "string"
          },
          "fps": {
            "default": 30,
            "title": "Fps",
            "type": "integer"
          },
          "max_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Duration Seconds"
          },
          "orientation": {
            "default": "landscape",
            "title": "Orientation",
            "type": "string"
          },
          "render_type": {
            "default": "slideshow",
            "title": "Render Type",
            "type": "string"
          },
          "seconds_per_image": {
            "default": 1.0,
            "title": "Seconds Per Image",
            "type": "number"
          }
        },
        "required": [
          "channel_archive",
          "brand_name"
        ],
        "title": "Body_batch_channel_api_batch_channel_post",
        "type": "object"
      },
      "Body_generate_shorts_api_shorts_generate_post": {
        "properties": {
          "audio": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio"
          },
          "fps": {
            "default": 30,
            "title": "Fps",
            "type": "integer"
          },
          "images": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          },
          "max_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Duration Seconds"
          },
          "orientation": {
            "default": "portrait",
            "title": "Orientation",
            "type": "string"
          },
          "output_name": {
            "default": "shorts.mp4",
            "title": "Output Name",
            "type": "string"
          },
          "seconds_per_image": {
            "default": 1.0,
            "title": "Seconds Per Image",
            "type": "number"
          },
          "shuffle": {
            "default": true,
            "title": "Shuffle",
            "type": "boolean"
          }
        },
        "required": [
          "images"
        ],
        "title": "Body_generate_shorts_api_shorts_generate_post",
        "type": "object"
      },
      "Body_generate_slideshow_api_slideshow_generate_post": {
        "properties": {
          "audio": {
            "format": "binary",
            "title": "Audio",
            "type": "string"
          },
          "fps": {
            "default": 30,
            "title": "Fps",
            "type": "integer"
          },
          "images": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          },
          "orientation": {
            "default": "landscape",
            "title": "Orientation",
            "type": "string"
          },
          "output_name": {
            "default": "slideshow.mp4",
            "title": "Output Name",
            "type": "string"
          }
        },
        "required": [
          "images",
          "audio"
        ],
        "title": "Body_generate_slideshow_api_slideshow_generate_post",
        "type": "object"
      },
      "Body_generate_video_api_lofi_generate_video_post": {
        "properties": {
          "audio": {
            "format": "binary",
            "title": "Audio",
            "type": "string"
          },
          "images": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          },
          "output_name": {
            "default": "output.mp4",
            "title": "Output Name",
            "type": "string"
          },
          "repeat_minutes": {
            "default": 60,
            "title": "Repeat Minutes",
            "type": "integer"
          }
        },
        "required": [
          "images",
          "audio"
        ],
        "title": "Body_generate_video_api_lofi_generate_video_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthPaths": {
        "properties": {
          "output": {
            "title": "Output",
            "type": "string"
          },
          "uploads": {
            "title": "Uploads",
            "type": "string"
          }
        },
        "required": [
          "uploads",
          "output"
        ],
        "title": "HealthPaths",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "paths": {
            "$ref": "#/components/schemas/HealthPaths"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "paths"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "JobCreateResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobCreateResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "output_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Filename"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "JobSummaryResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "output_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Filename"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "JobSummaryResponse",
        "type": "object"
      },
      "JobType": {
        "enum": [
          "lofi",
          "slideshow",
          "shorts",
          "batch"
        ],
        "title": "JobType",
        "type": "string"
      },
      "JobWebhookRequest": {
        "properties": {
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "JobWebhookRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Video generation API",
    "title": "FrameFusion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/batch/channel": {
      "post": {
        "operationId": "batch_channel_api_batch_channel_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_batch_channel_api_batch_channel_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Channel",
        "tags": [
          "Batch"
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummaryResponse"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/download": {
      "get": {
        "operationId": "download_job_output_api_jobs__job_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Job Output",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/webhook": {
      "post": {
        "operationId": "configure_job_webhook_api_jobs__job_id__webhook_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure Job Webhook",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/lofi/generate-video": {
      "post": {
        "operationId": "generate_video_api_lofi_generate_video_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_video_api_lofi_generate_video_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Video",
        "tags": [
          "Lofi"
        ]
      }
    },
    "/api/shorts/generate": {
      "post": {
        "operationId": "generate_shorts_api_shorts_generate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_shorts_api_shorts_generate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Shorts",
        "tags": [
          "Shorts"
        ]
      }
    },
    "/api/slideshow/generate": {
      "post": {
        "operationId": "generate_slideshow_api_slideshow_generate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_slideshow_api_slideshow_generate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Slideshow",
        "tags": [
          "Slideshow"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}